Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0df503e4-a968-4cc9-9850-f74fb1bb7a25). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388940789
Change-Id: Ib23978bf725a2a5d797479eeea7fd4bae9060499
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com