Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/410ec5a1-3fab-42cc-90d3-5d36c5ae1b8b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388949582
Change-Id: Idd6ae8420d9f327576de3b2c0de68a240f29f9f2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com