Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/40b681f0-a3b7-4f0b-8ef1-0e7ed5272629). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388955771
Bug: 388964271
Change-Id: Iac5d22a21bf1fcccb1b746708a944173184b5b13
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com