Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/59c36207-0ac1-47e9-80b6-e88780629052). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388955491
Bug: 388963855
Change-Id: Iffe479101c0e3c883ebcb522d99aae91f01dc9ec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com