Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ee7c9b2-cecb-4137-a55a-da3f9200fdf4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388975710
Change-Id: I8a28b4e9922da16232f0e6fb7312a3ce84311aba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com