Revert "Add mmd_native namespace to system properties"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/336c0ee8-469f-4611-99d9-5084456c893d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388974584
Change-Id: I9873fda0558c4e5e7419e08c4b9f2f1c446cc99a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com