[HSG] Allow class directive without policy
In some cases, we want to add a policy for a member (field/method) in
the policy file for a class that already has an annotation.

Previously, we had to always have some policy in the class line, which
could cause confusion because this will essentially override the
annotation, so even when you change the class annotation, that wouldn't
be picked up.

Now update the parser so that a class directive is allowed to omit
a policy.

Flag: EXEMPT host test change only
Bug: 292141694
Test: $ANDROID_BUILD_TOP/frameworks/base/ravenwood/scripts/run-ravenwood-tests.sh -s
Change-Id: Ifa37cd0ff74026bc1ba2c7e8599b159fa0a07412