Revert "Google RCS uses FTEU MO SMS for phone number verification"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5cbd40b7-8f45-46c1-916e-4ca40db81193). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388978684
Bug: 388978684
Change-Id: Ie6a301fd668db326c0e459f41bedebfb50cbf306
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com