Revert "Fix typo in CATEGORY_BROWSABLE comment."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/37820d52-8673-454f-88f2-b51c49247fb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9431940050b6485c23b9d7c890a9eddb368af225
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com