Fix top activity has wrong z-order
When we open more than 7 other applications and launch a portrait
Activity within a landscape application, on the core side, we set the
z-order of NewActivity to 11 and OldActivity to 10 (or potentially a
larger number, but the z-order of OldActivity is always smaller than
that of NewActivity).

In this scenario, on the shell side, within setupAnimHierarchy, the
z-order will be recalculated based on the algorithm in
calculateAnimLayer. During the screen rotation, typically four targets
are collected, which ensures the z-order of NewActivity is always 9.
However, since OldActivity has the type CHANGE, isIndependent returns
false, and as a result, its z-order is not updated. This causes
OldActivity to obscure NewActivity (10 > 9).

To address this issue, we should move the setLayer operation above the
isIndependent check to ensure that the z-order is set correctly.

Bug: 389591871
Flag: EXEMPT bugfix
Test: Manually test according to the problem reproduction steps

Change-Id: I947a29b21e9845a79ecd1252bf06e927cc3aab63