Revert "Ignore null action in BinaryTransparencyService$PackageU..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/204b4697-c9ff-4a0d-932a-37a66df32739). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cf02aa8db450b5f859916f3640fdf5fd251c7a4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com