Revert "locksettings: use ArrayUtils.zeroize()"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14caf98f-db38-42ed-8f01-82553168e646). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d39761a95d2bcee064670ed0efd680d32b189d2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com