Parallelize permissions file reads in SystemConfig
This change modifies SystemConfig so that files are opened and scanned
in parallel. The collections that were previously held for a
SystemConfig instance have been moved to a new data class ConfigInfo. A
ConfigInfo is filled in for each partition concurrently before being
merged back into mConfigInfo in order.

This change is intended to avoid the lock contention that occurs in
StartActivityManager due to the serial file processing taking too much
time. This results in a 33ms improvement in StartActivityManager (a
reduction of about 36%).

Test: build, flash, boot, verify latency improvements.
Bug: 383869534
Change-Id: Ic0acfa8b228d90136fa530f9605d8e1781dc284b