Don't notify about sdcard after reboot if notification is cleared
Notification about SD card is currently shown after the reboot
even if user has cleared the notification.

The notification will not be shown again after reboot if the
user has cleared the notification with this change.

The nofification will still be shown when the sdcard is
removed and inserted again even if the user has cleared the
notification.

Bug: 324088527
Test: Manual test
Flag: EXEMPT bug fix
Change-Id: I023fcb9a48120e7a7e9a4bc16606a1ffde8b0619