Revert "Remove Unused 'times' Import"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a549af48-b87f-400e-aa8e-8f032ca69cce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2b6edcff2e07732c63ba9bc52bfa06e97e5166be
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com