Set min_sdk_version of framework-connectivity-b to Baklava
The value of "current" is 35 at the moment. Since the build flag
that migrates VCN to Tethering module is enabled in March train,
on production Android V(35) devices, framework-connectivity-b will
be included in mainline and conflicts with the VCN framework code.
This patch solves this issue by using Baklava as the
min_sdk_version instead.

Bug: 390096003
Test: atest FrameworkVcnTests && atest CtsVcnTestCases
Flag: EXEMPT; bug fix
Change-Id: I427b2e0231e2ccdc870500399be9387036b14edc