Jarjar IndentingPrintWriter with android.net.connectivity prefix
This patch is to fix NoSuchMethodError thrown from VCN dumpsys when
VCN is in the Tethering mainline module.

Because VCN is in Tethering module, the VCN service will use a
jarjared IndentingPrintWriter. During dumpsys,
VcnManagementService#dump will create a IndentingPrintWriter
and pass it to VcnUnderlyingNetworkTemplate#dump. Then because
VcnUnderlyingNetworkTemplate fails to recognize the jarjared class, it
will throw NoSuchMethodError.

This patch fixes this issue by also jarjaring the IndentingPrintWriter
in VCN framework library. It is aligned with the end state where all
classes in the VCN framework code except APIs will be jarjared to
have a android.net.connectivity prefix

Bug: 390490391
Test: need to add new CTS
Flag: EXEMPT; low risk and small scope bug fix
Change-Id: I34be6dbf816c752f6ba0b77dc4ad98c8ad71e743