Revert "Stop statically including IndentingPrintWriter in VCN se..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cec2596f-8866-4cfb-b299-df25b8f0c847). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0702cc42b15b5c0abdb7028078c107aeee62ee3a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com