Revert "Store LockscreenCredential in non-movable byte array"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b268c8b-e14e-4db7-a696-ad2a55d9c010). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 391717789
Change-Id: I8e91826e304adfe9c448c555bab011a9828c1a3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com