Revert "Add flag android.nfc.nfc_hce_latency_events"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a025e102-2fdf-4e1e-99bb-29a45d52efcb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f96be016725403875b1228eaf2bd8cdc1fe9b9d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com