Integrate SDK system feature cache behind a flag
Add the SDK system feature cache to the hasSystemFeature query lookup.
This involves populating the initial cache array in system server, and
handing this as a compact int array (<1KB) to each process.

An alternative would be to stash this in the ApplicationSharedMemory,
which we may do as part of a follow-up, but that would require
additional complexity in routing build constants to the fixed native
array sizes. We can adopt that model in the future after validating
improvements with a trunk stable aconfig flag.

Flag: android.content.pm.cache_sdk_system_features
Bug: 326623529
Test: presubmit
Change-Id: I5549260cafff3ae08543c0c8d944e8cf396da224