Revert "Update ownership of the Desktop Experience flags."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c8d673dd-fbf4-4120-9843-51d5d2d34549). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4b6c96aaab3dd900dd727af7acbf416db2bcc65e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com