Revert "Jarjar IndentingPrintWriter with android.net.connectivit..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/111a1d0d-bc7b-43e6-981b-da859a6daef0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61cc3c1f26f285e4099e45afecd293f0b21da30a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com