Revert "Update ownership of the Desktop Experience flags."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9c74eb1-a7fc-41b4-b777-4207f8e9ab21). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2f8fedd3ec3b58c8b9b8254affcc1326c7654508
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com