Revert "Make Context#TETHERING_SERVICE public API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5fffed8-a50b-45dc-8a1e-99de796a12ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5e3a49bf42d569249c1f039f14ad83a5d337fa2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com