AudioRoutingHelper: set preferred device for audio media route
AudioRoutingHelper will allow to set
preferred device for a Media strategy and therefore set audio route to the chosen device. This is particularly necessary as when wired device is connected, the previously active device, visible in the Bluetooth Settings can not be set as preferred audio route. This change allows that so when the device is chosen in the settings the audio manager is informed that it's currently the preferred device for audio route.

Bug: 202832397
Flag: com.android.settingslib.flags.set_preferred_audio_route_for_bt_device_when_wired_connected
Test: atest SettingsLibRoboTests:com.android.settingslib.bluetooth.AudioRoutingHelperTest
Change-Id: Ic453b7e4fd3233faf74fa60fc9654a2c52b781b3