Revert "Make Context#TETHERING_SERVICE public API"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/adfdab8e-596e-49a9-9c98-5a1516d5eda6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8ce1e4fbf0f20f9e669cc457a896419fe4766f04
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com