Revert "Add Media Better Together team as OWNERS to its aconfig ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ad3a2f4-98ff-4d67-a00d-b01aa5fe1acc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7ddef39638007792f230d94b72f6be12ed48d91
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com