create a testlooper library for mainline tests
This change creates a testlooper library for mainline tests. This build
target has to have be configurated the sdk_vesion, since most mainline
module test should only build against certain stubs of the framework,
but not the implementation details of the framework.

Without setting the sdk_version of the library, if this library uses
some hidden methods in the framework, these mainline tests statically
links to it will also be enable to use the hidden methods.

Currently the problem is because the sdk_version is not set on the
library, the repacaging rules (jarjar rules) of the framework will be
propagated to the tests, and the tests will try to use the flags from
framework. Since the flags are hidden in framework, the test will fail.

Test: presubmit
Bug: 345013072
Change-Id: Ia6477f59e0eb72ae16cc03bd57592ac9c7672a14