audio: Fix A2DP volume issue with multiple outputs
Issue:
When the user controls the volume using the A2DP device's volume button
during multiple audio outputs, the volume index of the A2DP device does
not change. For example, when audio is output simultaneously from both
the speaker and an A2DP device, the Audio Framework's volume index
does not update even if the user adjusts the volume on the A2DP device.
The expected behavior is that the volume index for the A2DP device
should change.

Root Cause:
The Audio Framework selects a single device for volume updates.
During multiple audio outputs, it may select a device other than
the A2DP device. Then, the Audio Framework does not update the
volume index for the A2DP device.

Solution:
When the user controls the volume using the A2DP device button,
the AudioService ensures that the A2DP device is selected as the target
for volume adjustments.

Remark:
Similar issues may also occur with absolute volume devices.

Bug: 399307939
Test: manual - control the volume using the A2DP device's volume
button during multiple audio outputs

Change-Id: Ib8ad42278eacc2cd5969e78bd02cd43c99ec371d
Signed-off-by: Tomoki Yonezawa <Tomoki.Yonezawa@sony.com>