cec: box may not send <Standby> message [1/1]
PD#SWPL-201412
bug:383849602

Problem:
The box may fail to send a <Standby> message after sleep.
If a hotplug-out event occurs after the device goes to sleep
and a hotplug-in event happens later than the execution of the
queued DelayedStandbyRunnable, the flag indicating whether a
<Standby> message has been received is incorrectly updated.
As a result, the device will send an <Inactive Source> message
instead of a <Standby> message the next time.

Solution:
Add the power status condition check for the standby and
wake-up methods.

Verify:
adt4

Change-Id: I779b57f8721a4e7bb0304b9fa2686a056ea4a4fe
Signed-off-by: an.xi <an.xi@amlogic.com>