Revert "Add owners from Output Switcher team."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6fe85072-1018-4dce-b75c-dd77e26591d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 395479964
Change-Id: I463a714f3ee15d5dce3f8bb0a8a5c5f8e573eb56
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com