Revert "android_util_Binder: strict RefBase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c0915df-f416-4cb1-8fce-6ec5d26677cd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I088edba8f321b85e6a5f17c27e0ee730bb910126
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com