Revert "Add owners from Output Switcher team."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9c6ec7cb-ecbd-4979-9fb8-57ec24a48e10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0bf1705922be99f312b7b5674bdf2fec4d317278
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com