Revert "Deregistering mixes before register mixes in connectMixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d13107ab-8076-434f-9788-fbe036f5b35f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6006180d95f46d5650677f287a3d995a92eae2c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com