Add Screen and Lockscreen Status to Crash Files
Applications in the foreground can crash while the screen is off or when device is in the
lockscreen. Under these circumstances, the foreground field for these crash files will still show
as "Yes". This is not useful for developers who want to prioritize user perceivable crashes or
reproduce crashes occuring during screen off or on state. To help developers filter out these
scenarios, this CR adds screen and lockscreen information to crash files.

Test: manual
1. Created ANRs, watchdogs, crashes and verified that the screen headers were added.
2. Triggered crash while the screen was off/on and verified that the correct information showed up
   in the generated crash file.
3. Triggered crash while the device was in lockscreen and verified that the correct information
   showed up in the generated crash file.

Change-Id: Iddd77d43c789ef418196b06311d9fbfe04728e66
Signed-off-by: Aaron Huang <aahuang@amazon.com>