Revert "Deregistering mixes before register mixes in connectMixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f4872e3-a3e8-4cc3-ad4f-eabc4f88d9e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d2c050f569fd7552e857a219a729bb621b107b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com