Revert "Optimize string interning in ParsedComponentImpl seriali..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1652dc74-8356-4747-946d-652e6610a3ce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd5d3d331836e3d791ac1da3e8c30cdf72e9b412
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com