Revert "Add DeviceConfig overlay for mm. properties in ZramMaint..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae354b86-595a-4aec-af0d-accb93980a0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic047a9a938702054a78eeeae4ea5bdbd55b5adbf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com