Revert "Check isZramMaintenanceSupported on ZramMaintenance"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae354b86-595a-4aec-af0d-accb93980a0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4f2040dbfa4dbe10c2afd7e7cef90fb01d5370f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com