Revert "fw/base: work around mockito upgrade issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd2b6a57-f697-47c2-80e7-8b2d87726fa7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 398868342
Change-Id: Ibf634e6d232654c59f1e89435131226efe0b1cb0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com