Revert "fw/base: work around mockito upgrade issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ff61158-9ebb-4a1e-a0c5-f4b58a5fc295). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 399112039
Change-Id: I0df6edeadc11d2ee34926c129037d42d8493d4c4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com