Revert "Extend list of ciphers in conscrypt perftests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c32cd2d-0c0a-42b9-92d1-40115c1a87c8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b76b54bb0eb1fa720d1021af6e220cd3a757a54
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com