Revert "fw/base: work around mockito upgrade issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c31583b3-7719-45f9-bed8-15be710ec4eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 399295169
Change-Id: I281839f593481b9cee84078d76837059f0f2e9b6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com