AppCompatOverridesServiceTest: work around mocking issue
During testing, unregistering the receiver currently fails because that
is not overridden in MockContext, and then silently ignored because it
is called on the finalizer thread.

We leverage that to work around an issue where trying to unregister the
DeviceConfig listeners will interfere with stubbing in subsequent
threads, because that is not thread-safe.

By crashing earlier, the problematic unregistration is never invoked.
Note that this ultimately doesn't leak the listener registration,
because TestableDeviceConfig clears the listeners anyways.

Bug: 397997489
Test: atest AppCompatOverridesServiceTest
Flag: TEST_ONLY
Change-Id: Ic5ed380286df6cd41c9b1a504e237d320c37080c