Revert "Extend list of ciphers in conscrypt perftests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60e5dcbc-08fb-41e1-ac2c-b7fa74c8ff1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib676c07634b13365476e37a4299d396d61151399
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com