Revert "Extend list of ciphers in conscrypt perftests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b93d390b-91bc-4b1d-8b7f-7aeadaef295b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I130053a08f649a6bdc5f2938a9b53e4c104166f7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com