Remove private-stub-annotations-jar from *_current targets
The `private-stub-annotations-jar` contains definitions of various
`android.annotation` and `androidx.annotation` annotation classes that
have been modified to be package-private. It was created and added to
work around problems with the Android SDK in Android Studio by
providing definitions of the annotations without making those
annotations available as part of the SDK.

Unfortunately, they prevent any stubs created by Metalava that depend
on these from including annotations like `android.annotation.NonNull`.
That prevents `android.test.mock` stubs from including nullability
annotations.

This should be safe and not break Android Studio as these are only used
internally and not distributed as part of the Android SDK. That uses
the `*_exportable` versions which still include it.

Bug: 399365344
Test: m checkbuild
Change-Id: I56fb9d541ee22fefca416e0a432a9119e6906c89