Revert "AppCompatOverridesService: Remove finalizer and unregist..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a8d2971-f29d-4134-83c7-afea44b2a8fb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I95149761ab29aaba1d537dbe0864f99c3c0f9c48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com