Revert "Avoid NullPointerException if Location manager is disabled"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72461496-6b0f-4b9d-b93c-ddd04c944366). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id76fb5949b941323cdb471d8542854f1fe09752d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com