Revert "AppCompatOverridesService: Remove finalizer and unregist..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9ca62f8-2206-4738-8a51-421846d143b0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifed7be10dc2f7ca73e5fc4ee3ae71b9f676574f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com