Revert "Add install-time SDM signature verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55aa6146-5ee9-420a-8948-5b01bf126c12). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8caab86175162ee574e5cdadd2076acae1e79837
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com