Revert "Extend list of ciphers in conscrypt perftests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/020e34f2-5162-41ac-a4a6-9c3a2c474990). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6feda1262b57a338022b1ac9f95aa9f406828986
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com