Revert "AppCompatOverridesService: Remove finalizer and unregist..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/571662a7-7bad-438e-a153-3fd272ceb24d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d5b46d0b939b2de720b6ddd6a9d1792f51f4604
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com