Revert "Add install-time SDM signature verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7a8ed6af-22b1-4b89-9745-9eb297cfe39c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idfc5c5724c374460bc852e6c0cb816fa0c4f5d2e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com