Revert "Remove redundant code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba6d1c77-dcf4-481a-b99d-3735180002cf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9a6863a2e152ff24a0dd731760d625719e9773b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com