Revert "Remove redundant code."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5cb16481-7c68-4ce2-ada8-08ebbcbcc0d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I136462fe717a8c7277a1f94bc5183879a8bfcab7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com