Revert "Improve ANR debug store logging with more app lifecycle ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1507fe7c-1420-437a-a0db-1b437606f3f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7d9979a08c96a05e297f8efb184ed60bb1d72ee8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com