Revert "Add OWNERS for debug store"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1507fe7c-1420-437a-a0db-1b437606f3f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I448a0f05203117b84e1a2f43b0e29857f21f8bff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com