Revert "Add install-time SDM signature verification."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2ced2e5-141b-4184-acd3-4411ce24c0f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id46fa88b43e457caa95113cc7de357b9500a709b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com