Revert "Add system_ext partition to aconfig_flags.pb paths"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85249b62-c4a4-4aa3-b952-fa1bb31c051a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d6c177714965e9a6f0845e038298da4fbfccb6a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com