Revert "Add system_ext partition to aconfig_flags.pb paths"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c34a54c8-ed60-4e37-8745-723f2f717604). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id68e1b332ebdbc1af6d2c0aac1218315c21fd1db
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com