Revert "Add system_ext partition to aconfig_flags.pb paths"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cad100f9-4761-4707-b496-35b06eb7954e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d19c98e50b9e8426b34315dde3cee2565b0b349
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com