Revert "Add system_ext partition to aconfig_flags.pb paths"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/629515bb-dc39-4b59-a214-b6dce07cc473). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I43b520d6f9938c3b8bd2c95b131b93cb29ac37d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com