Revert "Allow enabling Adb logging dynamically"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6f8a4cff-6356-4e0a-885d-49bac8b4ce53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib32ba90374fbfb4def16906cd13480bd03dbbab6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com