Revert "fw/base: work around mockito upgrade issues"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94f35391-d2f1-45ff-9542-c92025e19a84). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 402265718
Change-Id: I606b2c3d4b06489c2ebf3c9f9cbfa73d7e39be6d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com