Revert "Adb wifi acquires multicast lock"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0d70556d-a7db-4844-b86e-1ef08012aec6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d29f9f7f4ff470f5cf1566945e68c0041ec963b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com