Checking userId in com.android.server.pm.ComputerEngine#resolveContentProvider
In the resolveContentProvider function, the entry checks whether the input parameter userId exists,
then the userId that may be carried in the uri is parsed,
and the userId in the uri is not checked for its existence.
As a result, in the queryProvider function,
packageState.getUserStateOrDefault(userId) obtains PackageUserStateInternal.DEFAULT,
thereby obtaining ApplicationInfo. If the userid in the uri is not created,
this is not in line with expectations.

In the resolveContentProvider function,
this patch adjusts the order of the judgment conditions for checking userId,
and checks whether it exists after obtaining the real userId.
This allows the expected results to be met in the cross-user query scenario.

Bug:401188957
Change-Id: I8509df5387dd0df604a4838e8fa1db8267d7e62d
Signed-off-by: huangdezhi@hihonor.com