Do not kill application not responding when being debugged
Summary: Do not kill an application that has stopped responding when debugging.

For debugging an application, the system promotes developers to explicitly mark the application for this procedure by means of a specific command. This change alters the behavior of the underlying method that forces the application to stop running. Now, it doesn't force the application to stop if the debugging is transient and doesn't require the application to wait for a debugger on restart. Instead, it marks the application record as 'debugging' to enable proper handling by the system during timeouts or unresponsive states. This enables the debugger to attach to the application.

Some system events are triggered frequently and failure with their broadcast can cause the debugged application to be terminated. Prevent the application from being terminated in case the broadcast transaction fails. Ignore the failure and don't propagate the exception further to assume that it can succeed in the future.

Test: Successful Build on master branch

Change-Id: I05c4aa99d8c03e5864b01f9cd730be2f4939cafd
Signed-off-by: Abhishek Gadewar <abhishekgadewar@fb.com>