Fix public volume not mount issue
In this case, that two user with keyguard and locked, when cold boot,
unlock user A, switch to user B, but not unlock, switch back to user A,
when remount, the volume user id and current user id is the same, so no
remount will happen, but the session already disconnected, so user A
cannot use that public volume.

Fix this by remove the conditon check of user id.

Test: Manual

Bug: 407792229
Change-Id: Id59118f72ea6bee0fc0b3731dd1471defb174d1d