statementservice: remove dodgy caching for domain verification
Summary: DomainVerifier is a singleton, and hence has the same lifetime
as StatementService itself. however, DomainVerifier caches all
certificate fingerprints it encounters by package name. this is
extremely dodgy, because you can uninstall and reinstall the
package with a different fingerprint, and the cache would return
the first one seen during the lifetime of StatementService.

remove the LruCache which has a size of 50! and force
DomainVerifier to fetch fresh signatures every time it wants to verify
something.


Test:

```
$ adb install HttpsDeeplinkHandler2.apk
Performing Streamed Install
Success
$ adb shell pm get-app-links com.example.httpsdeeplinkhandler
[...]
    Domain verification state:
      bryct-messing-around-public.s3.us-west-2.amazonaws.com: verified
$ adb uninstall com.example.httpsdeeplinkhandler
Success
$ adb install HttpsDeeplinkHandler1.apk
Performing Streamed Install
Success
$ adb shell pm get-app-links com.example.httpsdeeplinkhandler
[...]
    Domain verification state:
      bryct-messing-around-public.s3.us-west-2.amazonaws.com: legacy_failure

Change-Id: I37a59b375fde1cf6ae69d7acf9756b5c10e9e523
Signed-off-by: Pratik Kathalkar <pkathalk@meta.com>