Fix audiofocus trampling problem
Problem phenomenon: During listening to music, listen to multiple WeChat voices. After the first voice bar ends, the voice bar and music are suspended.

Cause of the problem: After the first voice bar ends, WeChat loses focus. When distributing focus to music, WeChat plays the second voice bar and applies for focus, which leads to focus trampling and both parties lose focus.

Solution: delay the focus distribution, and if an app continues to apply for focus within the delay time, it will not be distributed.

bug: 425083206

Change-Id: I5cadc35a296d04032f6ae3c929121e0a3fe0d85f
Signed-off-by: renyuxin <renyuxin@xiaomi.corp-partner.google.com>