Fix the listener in the incremental service.
Summary: The listener doesn't function as intended. The previous listener triggers for changes matching the given code or changes matching the package, always returning the listened code, not the changing code.

The end result is that logging will be disabled if there are any changes for a data loader. An example can be seen when the device is activated and a number of changes occur, resulting in the disable logging flow.

With this change, we listen to all changes for a given package name. This results in the listener being called with accurate codes, which we can then filter for.

Test: Successful Build on master branch

Change-Id: I44eca6cad70933eca9310a055e5839f8908a142b
Signed-off-by: Abdelrahman Daim <adaim@meta.com>