Fix removeSyncBarrier wakeup logic
Previously if the Looper was stuck on a barrier we would
find, remove, and recycle the barrier message, then check
if the barrier token to remove matches what was the first
message in the queue.

The bug is that we would recycle the message before checking
the token, hence override its arg1 field with 0.

This problem is fixed properly in an existing patch that was already
merged to the internal branch and will release to AOSP later.
This change is a minimal patch to fix reported regressions.

Bug: 422366962
Bug: 424123643
Flag: EXEMPT bugfix
Change-Id: I77d1e74dbf058bd902e949927f97cd7fc1fc7c2e