Add a test to check network connection
In certain live streaming video scenarios, if the network connection
is lost, the video player at the application layer will pause playback.
Unfortunately, it does not automatically resume when the network is restored.

Add a test to the media framework to check whether the network connection
is functioning normally. When the network is restored, the video should
continue playing using the previous URL.

Change-Id: I9e0c6e214eb669b301d7d9cc24ff3e6075066df9