Fix getting wrong binder proxy from cache.
This problem occurs because of multi-threaded operations.The binder thread calls the initServiceCache method to initialize the cache, and the main thread obtains data from the cache at the same time. However, the data in the cache has not been initialized yet, so dirty data is obtained, resulting in an exception.

Test: monkey test
Flag: EXEMPT bugfix
Bug: 433457366
Change-Id: I92b2ff15f15efc5ae414c6a8644207a94b4490f0