Allow boot time dynamic SystemConfig features
Features up until this point have been entirely build time. With this
change we utilize the same runtime sysprop-aware functionality that
static RROs utilize. Namely, the feature and unavailable-feature
SystemConfig elements now support requiredSystemPropertyName and
requiredSystemPropertyValue attributes which are used at boot time to
decide whether a feature is listed as available or unavailable.

This allows OEMs to turn features on or off based on runtime decisions.
Non-exhaustive list of examples:
  - A single build used for two different vehicle brands (ie Ford and
    Lincoln) where one brand has different features than the other
  - A feature which is to be turned on at a later date without requiring
    an OTA to do so.

One hesitation to this change might be that OEMs can now turn off
features later on. This could make it so applications installed from
Google Play with said feature as a requirement will be "orphaned" on a
device that no longer supports said feature. But this is not really any
different than the case where an OTA removes a feature that an app from
Google Play expected, so I do not feel it is a concern.

Test: manual for now, will add/update tests if patch is wanting to be
taken.

Change-Id: Iac0601da1968155c3caf4a5869a6ed531318ea2b